refactor eop.js
